Fixed incorrect dragging of floor/ceiling textures with "rotationfloor" / "rotationceiling" UDMF property set in GZDoom Visual Mode. Replaced all doubles with floats. Script Editor: Navigator drop-down now updates automatically. Navigator drop-down now also works for external files. Added Navigator drop-down support for Decorate and Modeldef.

+++ branches/GZDoomBuilder/Source/Core/Compilers/Compiler.cs
@@ -94,7 +94,7 @@ namespace CodeImp.DoomBuilder.Compilers
 			if(!isdisposed)
 			{
 				Exception deleteerror = null;
-				double starttime = General.Clock.CurrentTime;
+				float starttime = General.Clock.CurrentTime;
 				
 				do
 				{
