+++ branches/GZDoomBuilder/Source/Core/General/Clock.cs
@@ -22,6 +22,6 @@ namespace CodeImp.DoomBuilder
 	public class Clock
 	{
 		// This queries the system for the current time
-        public double CurrentTime { get { return Configuration.Timer.ElapsedMilliseconds; } }
+        public float CurrentTime { get { return Configuration.Timer.ElapsedMilliseconds; } }
 	}
 }
