+++ branches/GZDoomBuilder/Source/Plugins/GZDoomEditing/VisualModes/IVisualEventReceiver.cs
@@ -59,7 +59,7 @@ namespace CodeImp.DoomBuilder.GZDoomEditing
 		void OnTextureFloodfill();
 		void OnToggleUpperUnpegged();
 		void OnToggleLowerUnpegged();
-		void OnProcess(double deltatime);
+		void OnProcess(float deltatime);
 		void OnInsert();
 		void OnDelete();
